requirements.txt
requests
tqdm
pyguitarpro
mido